change posts

+++ src/controllers/post.controller.js
@@ -1,8 +1,9 @@
 const dateFilter = require('../_shared/query-date.helper');
 
-const { onSuccess, onCreated, onDeleted, onError } = require('../_shared/handlers/index');
+const { onSuccess, onCreated, onDeleted, onError, onUnathorized } = require('../_shared/handlers/index');
 
 const postService = require('../services/post.service');
+const commentService = require('../services/comment.service');
 
 class Controller {
     async create(ctx) {
@@ -13,6 +14,7 @@ class Controller {
 
             onCreated(ctx, created);
         } catch (err) {
+            console.log(err.stack);
             onError(ctx, err.message);
         }
     }
@@ -26,7 +28,8 @@ class Controller {
 
             onSuccess(ctx, created);
         } catch (err) {
-            onError(ctx, err);
+            console.log(err.stack);
+            onError(ctx, err.message);
         }
     }
 
@@ -57,6 +60,7 @@ class Controller {
 
             onSuccess(ctx, posts);
         } catch (err) {
+            console.log(err.stack);
             onError(ctx, err.message);
         }
     }
@@ -64,16 +68,18 @@ class Controller {
     async listComments(ctx) {
         try {
             const { query } = ctx.request;
+            const { id } = ctx.params;
 
             const pagination = {
                 skip: Number(query.skip) || 0,
                 limit: Number(query.limit) || 20
             }
 
-            const comments = await postService.list(filters, pagination);
+            const comments = await commentService.list({ post: id }, pagination);
 
             onSuccess(ctx, comments);
         } catch (err) {
+            console.log(err.stack);
             onError(ctx, err.message);
         }
     }
@@ -86,7 +92,8 @@ class Controller {
 
             onSuccess(ctx, user);
         } catch (err) {
-            onError(ctx, err);
+            console.log(err.stack);
+            onError(ctx, err.message);
         }
     }
 
@@ -94,11 +101,18 @@ class Controller {
         try {
             const { id } = ctx.params;
 
+            const post = await postService.getById(id);
+
+            if(post.author != ctx.userId) {
+                return onUnathorized(ctx, 'Only the post original author can remove!');
+            }
+
             await postService.remove(id);
 
             onDeleted(ctx);
         } catch (err) {
-            onError(ctx, err);
+            console.log(err.stack);
+            onError(ctx, err.message);
         }
     }
 }
